refactor: Refactored torrent content page and created page for settings and video streaming
