+++ lib/Pages/settings_screen.dart
@@ -1,0 +1,23 @@
+import 'package:flood_mobile/Constants/app_color.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_keyboard_visibility/flutter_keyboard_visibility.dart';
+
+class SettingsScreen extends StatefulWidget {
+  @override
+  _SettingsScreenState createState() => _SettingsScreenState();
+}
+
+class _SettingsScreenState extends State<SettingsScreen> {
+  @override
+  Widget build(BuildContext context) {
+    double hp = MediaQuery.of(context).size.height;
+    double wp = MediaQuery.of(context).size.width;
+    return KeyboardDismissOnTap(
+      child: Scaffold(
+        body: Container(
+          color: AppColor.primaryColor,
+        ),
+      ),
+    );
+  }
+}
